テストしやすいようにソースコードを実行部分とコンポーネントで分離
